Move to poetry and automated relases via Github UI
