--- .bumpversion.cfg
@@ -1,18 +1,0 @@
-[bumpversion]
-current_version = 6.1.2rc
-commit = True
-parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)((?P<release>.*))?
-serialize = 
-	{major}.{minor}.{patch}{release}
-	{major}.{minor}.{patch}
-tag_name = {new_version}
-
-[bumpversion:part:release]
-optional_value = production
-values = 
-	rc
-	production
-
-[bumpversion:file:setup.py]
-
-[bumpversion:file:src/pydocstyle/utils.py]
