+++ .github/workflows/pre-release.yml
@@ -1,14 +1,14 @@
 ---
-name: PyPI publish
+name: Test PyPI publish
 
 on:
   release:
-    types: [released]
+    types: [prereleased]
 
 jobs:
   build:
     runs-on: ubuntu-latest
-    environment: pypi-prod
+    environment: pypi-dev
     steps:
       - name: Checkout code
         uses: actions/checkout@v3
@@ -34,4 +34,4 @@ jobs:
         run: poetry build
 
       - name: Publish package
-        run: poetry publish -u __token__ -p ${{ secrets.PYPI_PASSWORD }}
+        run: poetry publish -r testpypi -u __token__ -p ${{ secrets.TEST_PYPI_PASSWORD }}
