+++ .github/workflows/test.yml
@@ -1,7 +1,12 @@
 name: Run tests
 
-on: [push, pull_request]
-
+on:
+  push:
+    branches:
+      - master
+  pull_request:
+    branches:
+      - master
 jobs:
   test-latest:
     runs-on: ${{ matrix.os }}
