+++ requirements/runtime.txt
@@ -1,2 +1,3 @@
-snowballstemmer==1.2.1
-toml==0.10.2
+snowballstemmer>=1.2.1
+toml>=0.10.2
+importlib-metadata<5.0.0,>=2.0.0; python_version < "3.8"
