+++ src/pydocstyle/_version.py
@@ -1,0 +1,13 @@
+import sys
+
+if sys.version_info[:2] >= (3, 8):
+    from importlib import metadata
+else:
+    import importlib_metadata as metadata  # pragma: no cover
+
+# Used to automatically set version number from github actions
+# as well as not break when being tested locally
+try:
+    __version__ = metadata.version(__package__)
+except metadata.PackageNotFoundError:  # pragma: no cover
+    __version__ = "0.0.0"
