+++ tox.ini
@@ -26,7 +26,7 @@ deps =
 [testenv:install]
 skip_install = True
 commands =
-    python setup.py bdist_wheel
+    pip wheel . -w dist --no-deps
     pip install --no-index --find-links=dist pydocstyle
     pydocstyle --help
 
@@ -64,4 +64,3 @@ commands = {[testenv:install]commands}
 [testenv:py310-install]
 skip_install = {[testenv:install]skip_install}
 commands = {[testenv:install]commands}
-
